pyproject.toml
[project]
name = "taas-server"
version = "0.1.0"
description = "Task-as-a-Service server for LLM agentic workflows"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # gRPC and Protocol Buffers
    "grpcio>=1.60.0",
    "grpcio-tools>=1.60.0",
    "protobuf>=4.25.0",
    
    # Async execution
    "celery>=5.3.0",
    "redis>=5.0.0",
    
    # Database
    "sqlalchemy>=2.0.0",
    "alembic>=1.13.0",
    
    # Artifact management
    "boto3>=1.34.0",  # S3-compatible storage
    
    # Validation and utilities
    "pydantic>=2.5.0",
    "jsonschema>=4.20.0",
    "pyyaml>=6.0.0",
    
    # LLM integration
    "openai>=1.10.0",
    "anthropic>=0.8.0",
    "google-generativeai>=0.3.0",
    
    # CLI and formatting
    "click>=8.1.0",
    "rich>=13.7.0",
    "python-dotenv>=1.0.0",
    
    # Logging
    "structlog>=24.1.0",
    
    # WebSocket support for agent API
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "websockets>=12.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "black>=24.0.0",
    "ruff>=0.1.0",
    "mypy>=1.8.0",
    "types-pyyaml",
    "types-redis",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "A", "C4", "DTZ", "T10", "DJ", "EM", "ISC", "ICN", "G", "PIE", "T20", "PYI", "PT", "Q", "RSE", "RET", "SLF", "SIM", "TID", "TCH", "ARG", "PTH", "ERA", "PD", "PGH", "PL", "TRY", "NPY", "RUF"]
ignore = ["E501", "TRY003", "PLR0913", "PLR2004"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --strict-markers --cov=taas_server --cov=llm_agent --cov-report=term-missing"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
strict_equality = true

[project.scripts]
taas-server = "taas_server.server:main"
taas-agent = "llm_agent.cli.agent_cli:main"
